// frontend-svelte/src/lib/payment/interfaces/PaymentProvider.ts
export interface WebhookEvent {
  id: string
  type: string
  data: {
    object: Record<string, unknown>
    previous_attributes?: Record<string, unknown>
  }
  created: Date
}